Change cents amount to be float not int

+++ checkout/views.py
@@ -42,7 +42,7 @@ def process_payment(request):
         with transaction.atomic():
             # Convert amount from dollars to cents for Stripe
             amount_in_dollars = float(data['amount'])
-            amount_in_cents = int(amount_in_dollars * 100)
+            amount_in_cents = float(amount_in_dollars * 100)
             subscription_id = data['subscription_id']
 
             # Create a Stripe charge
